reduce js bundle by 178 KB (6.6%)
