+++ datahub-web-react/src/app/entityV2/shared/tabs/Dataset/Validations/assertion/profile/actions/ExternalUrlAction.tsx
@@ -1,4 +1,4 @@
-import { LaunchOutlined } from '@mui/icons-material';
+import { ArrowUpRight } from '@phosphor-icons/react/dist/csr/ArrowUpRight';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
@@ -8,11 +8,8 @@ import { ActionItem } from '@app/entityV2/shared/tabs/Dataset/Validations/assert
 
 import { Assertion, AssertionRunStatus, EntityType } from '@types';
 
-const StyledLaunchOutlined = styled(LaunchOutlined)`
-    && {
-        display: flex;
-        font-size: 16px;
-    }
+const StyledLaunchOutlined = styled(ArrowUpRight)`
+    display: flex;
 `;
 
 type Props = {
@@ -52,7 +49,7 @@ export const ExternalUrlAction = ({ assertion, isExpandedView = false }: Props) 
             key="external-url"
             tip={t('action.viewInPlatform', { platformName })}
             onClick={handleRedirect}
-            icon={<StyledLaunchOutlined />}
+            icon={<StyledLaunchOutlined size={16} />}
             isExpandedView={isExpandedView}
             actionName={t('action.viewExternalPlatform')}
         />
