+++ datahub-web-react/src/app/sharedV2/forms/multiStepForm/MultiStepFormContext.tsx
@@ -1,6 +1,5 @@
 import deepEqual from 'fast-deep-equal';
 import React, { useCallback, useMemo, useState } from 'react';
-import { deepMerge } from 'remirror';
 
 import {
     MultiStepFormContextType,
@@ -9,6 +8,21 @@ import {
     Step,
     StepKey,
 } from '@app/sharedV2/forms/multiStepForm/types';
+
+function deepMerge<T extends Record<string, unknown>>(target: T, source: Partial<T>): T {
+    const result = { ...target };
+    Object.keys(source).forEach((k) => {
+        const key = k as keyof T;
+        const srcVal = source[key];
+        const tgtVal = target[key];
+        if (srcVal && typeof srcVal === 'object' && !Array.isArray(srcVal) && tgtVal && typeof tgtVal === 'object') {
+            result[key] = deepMerge(tgtVal as Record<string, unknown>, srcVal as Record<string, unknown>) as T[keyof T];
+        } else if (srcVal !== undefined) {
+            result[key] = srcVal as T[keyof T];
+        }
+    });
+    return result;
+}
 
 const MultiStepContext = React.createContext<MultiStepFormContextType<any, any>>({
     state: {},
@@ -76,7 +90,10 @@ export function MultiStepFormProvider<TState, TSubmitOptions = any>({
     const totalSteps = useMemo(() => steps.length, [steps]);
 
     const updateState = useCallback((newState: Partial<TState>) => {
-        setState((currentState) => deepMerge(currentState ?? {}, newState));
+        setState((currentState) => {
+            const base = (currentState ?? {}) as Record<string, unknown>;
+            return deepMerge(base, newState as Record<string, unknown>) as TState;
+        });
     }, []);
 
     const [currentStepIndex, setCurrentStepIndex] = useState<number>(initialStepIndex);
